+++ src/flockwave/server/ext/mavlink/extension.py
@@ -20,6 +20,7 @@ from .enums import MAVSeverity
 from .errors import InvalidSigningKeyError
 from .network import MAVLinkNetwork
 from .rssi import RSSIMode
+from .rtk import RTKCorrectionPacketSignalManager
 from .tasks import check_uavs_alive
 from .types import (
     MAVLinkMessage,
@@ -58,6 +59,9 @@ class MAVLinkDronesExtension(UAVExtension[MAVLinkDriver]):
     log: Logger
 
     _networks: dict[str, MAVLinkNetwork]
+    """Dictionary mapping network IDs to the MAVLink networks managed by this
+    extension.
+    """
 
     def __init__(self):
         super().__init__()
@@ -101,7 +105,6 @@ class MAVLinkDronesExtension(UAVExtension[MAVLinkDriver]):
         # Get a handle to the signals extension that we will need
         signals = app.import_api("signals")
         status_summary_signal = signals.get("mavlink:status_summary")
-        rtk_packet_fragments_signal = signals.get("mavlink:rtk_fragments")
 
         # Create self._uavs only here and not in the constructor; this is to
         # ensure that we cannot accidentally register a UAV when the extension
@@ -114,10 +117,13 @@ class MAVLinkDronesExtension(UAVExtension[MAVLinkDriver]):
             "driver": self._driver,
             "log": self.log,
             "register_uav": self._register_uav,
-            "rtk_packet_fragments_signal": rtk_packet_fragments_signal,
             "supervisor": app.supervise,
             "use_connection": app.connection_registry.use,
         }
+
+        # Create an object responsible for distributing RTK correction packets
+        # to other extensions that are interested in them
+        rtk_correction_packet_signal_manager = RTKCorrectionPacketSignalManager()
 
         # Create a cleanup context and run the extension
         with ExitStack() as stack:
@@ -134,6 +140,13 @@ class MAVLinkDronesExtension(UAVExtension[MAVLinkDriver]):
                         "show:lights_updated": self._on_show_light_configuration_changed,
                     }
                 )
+            )
+
+            # Connect the RTK correction packet signal manager to the signals API
+            # so it knows which signal to dispatch when a new RTK correction
+            # packet is to be forwarded to other extensions
+            stack.enter_context(
+                rtk_correction_packet_signal_manager.use(signals, log=self.log)
             )
 
             # Forward the current start configuration for the drones in this network.
@@ -307,6 +320,7 @@ class MAVLinkDronesExtension(UAVExtension[MAVLinkDriver]):
         if not self._networks:
             return
 
+        # Forward the packet to all networks
         for name, network in self._networks.items():
             try:
                 network.enqueue_rtk_correction_packet(packet)